Fix build & bottom temps on chart (needs testing)
