+++ packages/website/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "start": "cross-env EXTEND_ESLINT=true react-scripts start",
-    "build": "react-scripts build",
+    "build": "cross-env EXTEND_ESLINT=true react-scripts build",
     "test": "react-scripts --expose-gc test --logHeapUsage --maxWorkers=2",
     "eject": "react-scripts eject",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx ./src --ignore-path .gitignore",
